+++ app/models/producto.model.js
@@ -1,16 +1,16 @@
 module.exports = (sequelize, Sequelize) => {
-    const Producto = sequelize.define("Producto", {
-        id: {
-            type: Sequelize.BIGINT,
-            primaryKey: true,
-            autoIncrement: true,
-        },
-        nombre: {
-            type: Sequelize.STRING,
-        },
-        precio_venta: {
-            type: Sequelize.BIGINT,
-        },
-    });
-    return Producto;
+  const Producto = sequelize.define("Producto", {
+    id: {
+      type: Sequelize.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    nombre: {
+      type: Sequelize.STRING,
+    },
+    precio_venta: {
+      type: Sequelize.BIGINT,
+    },
+  });
+  return Producto;
 };
